+++ tests/test_coding/test_iterators.py
@@ -13,9 +13,8 @@ class TestIterators(unittest.TestCase):
         g = hint_with_length(generator(), 1000)
         self.assertEqual(g.__length_hint__(), 1000)
 
-    @unittest.skipUnless(sys.implementation.name == 'cpython', 'Not CPython')
+    @unittest.skipUnless(sys.implementation.name == 'cpython', 'Not CPython, this needs deterministic GC')
     def test_self_closing_generator(self):
-
         a = {'done': False}
 
         def generator():
@@ -24,6 +23,21 @@ class TestIterators(unittest.TestCase):
             a['done'] = True
 
         for i in SelfClosingGenerator(generator()):
+            if i == 2:
+                break
+
+        self.assertTrue(a['done'])
+
+    @unittest.skipUnless(sys.implementation.name == 'cpython', 'Not CPython, this needs deterministic GC')
+    def test_self_closing_generator_function(self):
+        a = {'done': False}
+
+        def generator():
+            for i in range(5):
+                yield i
+            a['done'] = True
+
+        for i in SelfClosingGenerator(generator)():
             if i == 2:
                 break