+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 # v2.7.34
 
+* enabled `SelfClosingGenerator` to accept generator-generating functions as well
+
+